Deprecated old Core configuration method.
This shouldn't have been added - the component registration system can
provide proper configuration instead of relying on a static
`configureWithApp:` call. Many of the SDKs relied on singletons being
available there, which wasn't necessary. Instead, this should happen
in the component creation block and use the `instantiationTiming`
parameter on `registerInternalLibrary`